move hardcoded data to new data directory and update imports
