+++ data/connections/source-connection-logos.js
